+++ cloud-server-provider/user-server/user-server-provider-core/src/main/java/com/galaxy/microservice/user/core/service/impl/PermissionServiceImpl.java
@@ -16,7 +16,7 @@ public class PermissionServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> i
     private SysMenuMapper menuMapper;
 
     @Override
-    public List<SysMenu> getPermissionsByRoleId(Integer roleId) {
+    public List<SysMenu> getPermissionsByRoleId(Long roleId) {
         return menuMapper.getPermissionsByRoleId(roleId);
     }
 }
